+++ HISTORY.rst
@@ -7,6 +7,7 @@ History
 
 * Add python-logstash as requirement (ericof).
 * BODY-60: Fix scale url generation (ericof).
+* BODY-62: Add listing and summary serialization to objects (ericof).
 
 1.1.2 (2016-10-04)
 ------------------